+++ source/mango/jpeg/jpeg_process_func.hpp
@@ -4,7 +4,7 @@
 */
 
 #ifdef FUNCTION_GENERIC
-void FUNCTION_GENERIC(u8* dest, int stride, const s16* data, ProcessState* state, int width, int height)
+void FUNCTION_GENERIC(u8* dest, size_t stride, const s16* data, ProcessState* state, int width, int height)
 {
     u8 result[JPEG_MAX_SAMPLES_IN_MCU];
 
@@ -73,7 +73,7 @@ void FUNCTION_GENERIC(u8* dest, int stride, const s16* data, ProcessState* state
 #endif
 
 #ifdef FUNCTION_YCBCR_8x8
-void FUNCTION_YCBCR_8x8(u8* dest, int stride, const s16* data, ProcessState* state, int width, int height)
+void FUNCTION_YCBCR_8x8(u8* dest, size_t stride, const s16* data, ProcessState* state, int width, int height)
 {
     u8 result[64 * 3];
 
@@ -109,7 +109,7 @@ void FUNCTION_YCBCR_8x8(u8* dest, int stride, const s16* data, ProcessState* sta
 #endif
 
 #ifdef FUNCTION_YCBCR_8x16
-void FUNCTION_YCBCR_8x16(u8* dest, int stride, const s16* data, ProcessState* state, int width, int height)
+void FUNCTION_YCBCR_8x16(u8* dest, size_t stride, const s16* data, ProcessState* state, int width, int height)
 {
     u8 result[64 * 4];
 
@@ -149,7 +149,7 @@ void FUNCTION_YCBCR_8x16(u8* dest, int stride, const s16* data, ProcessState* st
 #endif
 
 #ifdef FUNCTION_YCBCR_16x8
-void FUNCTION_YCBCR_16x8(u8* dest, int stride, const s16* data, ProcessState* state, int width, int height)
+void FUNCTION_YCBCR_16x8(u8* dest, size_t stride, const s16* data, ProcessState* state, int width, int height)
 {
     u8 result[64 * 4];
 
@@ -200,7 +200,7 @@ void FUNCTION_YCBCR_16x8(u8* dest, int stride, const s16* data, ProcessState* st
 #endif
 
 #ifdef FUNCTION_YCBCR_16x16
-void FUNCTION_YCBCR_16x16(u8* dest, int stride, const s16* data, ProcessState* state, int width, int height)
+void FUNCTION_YCBCR_16x16(u8* dest, size_t stride, const s16* data, ProcessState* state, int width, int height)
 {
     u8 result[64 * 6];
